feat(page-info): add page info tab to default page blueprint to display creation and modification dates
A new tab called "Info" has been added to the default page blueprint. This tab displays the creation and last modification dates of the page. Two new fields have been added to the page-info.yml blueprint file to display this information. The "dateCreatedInfo" field displays the creation date of the page, while the "dateModifiedInfo" field displays the last modification date of the page.