🎨 refactor(blueprints, snippets, plugins): replace 'background-color-select' with 'brand-color-select' for better color management
The changes were made to improve the color management in the application. Instead of using a generic 'background-color-select', a 'brand-color-select' is now used. This allows for a more consistent and brand-aligned color selection across the application. The changes also include the creation of a new 'brand-color-select' field and the modification of several files to use this new field. This will make it easier to manage and update the brand colors in the future.